requirements.txt
Flask
flask-restful
flask-pymongo
flask-bcrypt
flask-jwt-extended
flask_marshmallow
python-dotenv